Add onNext support for ranges.
If onNext is called with an input range, the range will be iterated and the underlying single-value onNext method will be called with each of those values.  In effect, calling onNext with a single-element range will pass one call to the Observer's implementation of onNext; calling onNext with an n-element range will pass n calls to the underlying onNext.